Page vs Post Differientiation
I think that posts and pages should be a little more different and one aspect that I delt with is getting rid of the date functionality for pages. I really don't want people to know when I decided to make a page because it is static, but I do want people to know when I created a post because it shows a progression over time (dynamic).